# config/swagger.yml
swagger: '2.0'
info:
  version: v1
  title: wecard
  description: The wecard APIs based on Swagger and built using Node.js
host: api.wecard.com
schemes:
  - http
  - https
basePath: /api
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
security:
  - Bearer: []
consumes:
  - application/json
  - text/json
produces:
  - 'application/json, text/json'
  - application/json
  - text/json
paths:
...
  /cardcaseItem:
    get:
      tags:
        - CardcaseItem
      summary: Get all CardcaseItems by Cardcase id
      operationId: cardcaseItem_getByCardcaseId
      parameters:
        - name: cardcaseId
          type: string
          in: query
          required: true
          description: The Cardcase id
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/CardcaseItem'
    post:
      tags:
        - CardcaseItem
      summary: Create a CardcaseItem in Cardcase
      operationId: cardcaseItem_create
      parameters:
        - $ref: '#/parameters/CardcaseItemCreate'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/CardcaseItem'
...
  /group:
    get:
      tags:
        - Group
      summary: Get group by WX open group id
      operationId: group_getByWXOpenGid
      parameters:
        - name: wxGroupId
          type: string
          in: query
          required: true
          description: The WX open group id
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Group'
    put:
      tags:
        - Group
      summary: Update group
      operationId: group_update
      parameters:
        - $ref: '#/parameters/GroupUpdate'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Group'
...
  /groupMember:
    get:
      tags:
        - GroupMember
      summary: Get all GroupMembers by Group id
      operationId: groupMemeber_getbyGroupId
      parameters:
        - name: groupId
          type: string
          in: query
          required: false
          description: The Group id
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/GroupMember'
    put:
      tags:
        - GroupMember
      summary: Update self group member info
      operationId: groupMemeber_updateSelfInfo
      parameters:
        - $ref: '#/parameters/GroupMemberUpdate'
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/GroupMember'
  /accessrequest:
    get:
      tags:
        - AccessRequest
      summary: Get all AccessRequest sent to current user
      operationId: accessreqeust_get
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/AccessRequest'
    post:
      tags:
        - AccessRequest
      summary: Create an AccessRequest
      operationId: accessreqeust_create
      parameters:
        - $ref: '#/parameters/AccessRequestCreate'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/AccessRequest'
  '/accessrequest/{id}/sign':
    put:
      tags:
        - AccessRequest
      summary: Accept or refuse the AccessRequest by id
      operationId: accessRequest_signeById
      parameters:
        - $ref: '#/parameters/EntityId'
        - name: operation
          in: query
          description: The operation to be performed (accept|reject)
          required: true
          type: string
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/AccessRequest'
...
definitions:
...
  CardcaseItem:
    type: object
    properties:
      id:
        type: string
      userId:
        type: string
      itemId:
        type: string
      itemType:
        type: string
      cardcaseId:
        type: string
      name:
        type: string
      thumbnail:
        type: string
  Group:
    type: object
    properties:
      id:
        type: string
      wxGroupId:
        type: string
      name:
        type: string
  GroupMember:
    type: object
    properties:
      id:
        type: string
      groupId:
        type: string
      userId:
        type: string
      cardId:
        type: string
      hidden:
        type: boolean
  AccessRequest:
    type: object
    properties:
      id:
        type: string
      toNamecardId:
        type: string
      toUserId:
        type: string
      toUserName:
        type: string
      fromUserId:
        type: string
      fromNamecardId:
        type: string
      fromUserName:
        type: string
      status:
        type: string
parameters:
  EntityId:
    required: true
    name: id
    type: string
    in: path
    description: The entity id
...
  CardcaseItemCreate:
    name: body
    in: body
    required: true
    description: The Cardcase
    schema:
      type: object
      properties:
        itemId:
          type: string
        itemType:
          type: string
        cardcaseId:
          type: string
  GroupUpdate:
    name: body
    in: body
    required: true
    description: The Group
    schema:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
  GroupMemberUpdate:
    name: body
    in: body
    required: true
    description: The GroupMember
    schema:
      type: object
      properties:
        id:
          type: string
        cardId:
          type: string
        hidden:
          type: boolean
      required:
        - id
  AccessRequestCreate:
    name: body
    in: body
    required: true
    description: The AccessRequest
    schema:
      type: object
      properties:
        toNamecardId:
          type: string
        fromNamecardId:
          type: string